+++ src/operations/cat.png
